requirements.txt
Flask==3.0.0
Pillow==11.0.0
requests==2.31.0
beautifulsoup4==4.12.3
PyGithub==2.1.1
python-dotenv==1.0.0